@@ -1,35 +1,37 @@
 name: wassword
 description: Wassword - Wonderful Password Generator
 
-version: 2.4.0+24
+version: 3.0.0+25
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=3.1.0 <4.0.0'
 
 dependencies:
-  equatable: ^2.0.5
+  flex_seed_scheme: ^3.0.0
   flutter:
     sdk: flutter
-  flutter_bloc: ^8.1.1
-  fluttertoast: ^8.1.1
-  go_router: ^5.1.1
-  google_fonts: ^3.0.1
-  share_plus: ^6.2.0
-  url_launcher: ^6.1.6
-  window_manager: ^0.2.7
+  flutter_riverpod: ^2.5.1
+  flutter_styled_toast: ^2.2.1
+  go_router: ^14.2.0
+  package_info_plus: ^8.0.0
+  riverpod_annotation: ^2.3.5
+  share_plus: ^9.0.0
+  universal_platform: ^1.1.0
+  url_launcher: ^6.3.0
+  window_manager: ^0.3.9
 
 dev_dependencies:
-  flutter_launcher_icons: ^0.9.2
-  flutter_native_splash: ^2.2.0+1
-  flutter_lints: ^2.0.1
+  build_runner: ^2.4.11
+  custom_lint: ^0.6.4
+  flutter_launcher_icons: ^0.13.1
+  flutter_lints: ^4.0.0
+  flutter_native_splash: ^2.4.0
   flutter_test:
     sdk: flutter
+  riverpod_generator: ^2.4.0
+  riverpod_lint: ^2.3.10
 
 flutter:
-
-  # The following line ensures that the Material Icons font is
-  # included with your application, so that you can use the icons in
-  # the material Icons class.
   uses-material-design: true
 
   assets:
