@@ -1,18 +1,21 @@
+import 'package:flex_seed_scheme/flex_seed_scheme.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
-import 'package:google_fonts/google_fonts.dart';
 import 'package:wassword/pages/about_page.dart';
-import 'package:wassword/pages/home_page.dart';
-import 'package:wassword/styles/colors.dart' as mcolors;
+import 'package:wassword/pages/home/home_page.dart';
+import 'package:wassword/provider/brightness_provider.dart';
 
-class App extends StatelessWidget {
-  App({Key? key}) : super(key: key);
+class App extends ConsumerWidget {
+  App({super.key});
 
-  final _router = GoRouter(
+  final router = GoRouter(
     routes: [
       GoRoute(
         path: '/',
-        builder: (context, state) => const HomePage(),
+        builder: (context, state) {
+          return const HomePage();
+        },
       ),
       GoRoute(
         path: '/about',
@@ -22,25 +25,22 @@ class App extends StatelessWidget {
   );
 
   @override
-  Widget build(BuildContext context) => MaterialApp.router(
-        debugShowCheckedModeBanner: false,
-        routeInformationProvider: _router.routeInformationProvider,
-        routeInformationParser: _router.routeInformationParser,
-        routerDelegate: _router.routerDelegate,
-        theme: ThemeData(
-          useMaterial3: true,
-          textTheme: GoogleFonts.robotoTextTheme(
-            Theme.of(context).textTheme,
+  Widget build(BuildContext context, WidgetRef ref) {
+    return MaterialApp.router(
+      debugShowCheckedModeBanner: false,
+      routeInformationProvider: router.routeInformationProvider,
+      routeInformationParser: router.routeInformationParser,
+      routerDelegate: router.routerDelegate,
+      theme: ThemeData(
+        useMaterial3: true,
+        colorScheme: SeedColorScheme.fromSeeds(
+          primaryKey: Colors.deepPurple,
+          brightness: ref.watch(brightnessProvider),
+          tones: FlexTones.material(
+            ref.watch(brightnessProvider),
           ),
-          appBarTheme: AppBarTheme(
-            backgroundColor: mcolors.backgroundView,
-            centerTitle: true,
-            elevation: 0,
-          ),
-          brightness: Brightness.dark,
-          visualDensity: VisualDensity.adaptivePlatformDensity,
-          backgroundColor: mcolors.backgroundView,
-          scaffoldBackgroundColor: mcolors.backgroundView,
         ),
-      );
+      ),
+    );
+  }
 }
