@@ -2,5 +2,4 @@ package com.albertobonacina.wassword
 
 import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity: FlutterActivity() {
-}
+class MainActivity: FlutterActivity()
