@@ -1,50 +1,31 @@
 // ignore_for_file: depend_on_referenced_packages
 
-import 'dart:async';
-import 'dart:developer';
-import 'dart:io';
-
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_web_plugins/url_strategy.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:universal_platform/universal_platform.dart';
 import 'package:wassword/app/app.dart';
-import 'package:wassword/app/app_bloc_observer.dart';
 import 'package:window_manager/window_manager.dart';
-import 'package:wassword/styles/colors.dart' as mcolors;
-import 'package:flutter/foundation.dart' show kIsWeb;
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
   usePathUrlStrategy();
 
-  if (!kIsWeb &&
-      !Platform.isAndroid &&
-      !Platform.isIOS &&
-      (Platform.isWindows || Platform.isLinux || Platform.isMacOS)) {
+  WindowOptions windowOptions = const WindowOptions(
+    //Pixel 7a form factor (2400x1080 px)
+    size: Size(376, 835),
+  );
+
+  if (!UniversalPlatform.isWeb &&
+      !UniversalPlatform.isAndroid &&
+      !UniversalPlatform.isIOS &&
+      UniversalPlatform.isDesktop) {
     await windowManager.ensureInitialized();
-    windowManager.waitUntilReadyToShow().then((_) async {
-      await windowManager.setSize(const Size(385, 835));
+    windowManager.waitUntilReadyToShow(windowOptions, () async {
+      await windowManager.show();
+      await windowManager.focus();
     });
   }
 
-  SystemChrome.setSystemUIOverlayStyle(
-    SystemUiOverlayStyle(
-        systemNavigationBarColor:
-            mcolors.backgroundView, // navigation bar color
-        statusBarColor: mcolors.backgroundView, //top bar color
-        statusBarIconBrightness: Brightness.light, // status bar icons' color
-        statusBarBrightness: Brightness.dark),
-  );
-
-  Bloc.observer = AppBlocObserver();
-
-  FlutterError.onError = (details) {
-    log(details.exceptionAsString(), stackTrace: details.stack);
-  };
-
-  runZonedGuarded(
-    () => runApp(App()),
-    (error, stackTrace) => log(error.toString(), stackTrace: stackTrace),
-  );
+  runApp(ProviderScope(child: App()));
 }
