@@ -1,10 +1,31 @@
-double roundedCorner = 20;
+import 'package:flutter/material.dart';
 
-double paddingButtonBig = 20;
-double paddingButton = 16;
-double paddingHorizontal = 16;
+class Dimens {
+  // add a private constructor to prevent this class being instantiated
+  // e.g. invoke `Dimens()` accidentally
+  Dimens._();
 
-double defaultSpace = 16;
-double hugeSpace = 32;
+  static const double hugeRoundedCorner = 32;
+  static const double mainRoundedCorner = 16;
+  static const double smallRoundedCorner = 8;
+  static const double tinyRoundedCorner = 4;
 
-double heightSlider = 16;
+  static const double hugePadding = 20;
+  static const double mainPadding = 16;
+  static const double smallPadding = 8;
+  static const double tinyPadding = 4;
+
+  static const double hugeMargin = 20;
+  static const double mainMargin = 16;
+  static const double smallMargin = 8;
+  static const double tinyMargin = 4;
+
+  static const double hugeSpace = 32;
+  static const double mainSpace = 16;
+  static const double smallSpace = 8;
+
+  static const double heightSlider = 16;
+
+  static BorderRadius mainBorderRadius =
+      BorderRadius.circular(mainRoundedCorner);
+}
